Remove conv function from magnetar/fit_stats.py
Remove conv function to calculate Gelman-Rubin statistic from
magnetar/fit_stats.py. It is not correct to use the Gelman-Rubin
statistic to test convergence of multiple chains from the same
emcee sample as they are not independent! Therefore remove the
function. The output of this function was never reported in papers
anyway and so it is not critical to reproducibility.